refactor(debug): clean up files
